fix(model): give Bedrock an observed path instead of silence (#1132)
Every other adapter overrides `complete_observed_ref` and threads the
observer through its SSE parse loop. `bedrock.rs` inherited the trait
default, which discards the observer entirely — so a host watching the
observed path (the TUI's live rendering, stella-serve forwarding deltas
to a browser) saw nothing at all until the whole call returned, and
rendered as if the process had hung.

Converse is genuinely not a stream, so real token-level preview is out
of reach without the `application/vnd.amazon.eventstream` decoder. But
"one delta carrying the whole answer" is strictly better than zero, and
the `text_delta` contract is explicitly best-effort, so the single
terminal emission is honest rather than a masked bug.

Deliberately not announcing tool calls: `tool_call_streamed` starts
speculative execution, and with the whole body already parsed a call
announced here would beat dispatch by microseconds while paying a spawn
and emitting a discard. The module docs now say so rather than leaving
the omission to be rediscovered.

Tests move to `bedrock/tests.rs`, matching anthropic/gemini/zai — the
file-size ratchet's own guidance (split, don't raise the ceiling), which
retires bedrock.rs's grandfathered baseline entry.
